feat(cli): rewrite stc-builder with better API
